Update manual-sketching.scrbl (#31)
* Update manual-sketching.scrbl

Updated frame-rate examples to resemble the processing examples (https://processing.org/reference/frameRate.html) more accurately. In Processing, the frame rate is printed to the console. Also, the on-screen frame rate was invisible against the white background.

* Update manual-sketching.scrbl

That frame-count change was a mistake, sorry. I meant to only change the frame-rate stuff.